server/requirements_basic.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
openai>=1.0.0
requests>=2.31.0







